requirements.txt
httpx
python-dotenv
openai
deepdiff
